+++ ChangeLog.md
@@ -13,9 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `e3nn.IrrepsArray.convert` is now private (`e3nn.IrrepsArray._convert`) because it's recommended to other methods instead.
 - **breaking change** use `input.regroup()` in `e3nn.Linear` which can change the structure of the parameters dictionary.
 - **breaking change** `regroup_output` is `True` by default in `e3nn.tensor_product` and `e3nn.tensor_square`.
+- To facilitate debugging, if not `key` is provided to `e3nn.normal` it will use the hash of the irreps.
 
 ### Removed
-- Deprecated `e3nn.TensorSquare`
+- Deprecate `e3nn.TensorSquare`
 
 ## [0.11.1] - 2022-11-13
 ### Added